test(event): consolidate gap-fill tests into event_test.go
Move the gap-fill unit tests and their helpers (captureStderr, holeServer,
swapGapFillNow, swapGapFillVars) out of the separate gapfill_test.go and into
event_test.go, then delete gapfill_test.go. The functions under test
(applyChunk, giveUpGap, drainRemainingChunks, recoverSkippedChunks) all live in
event.go, so the tests belong beside the other event.go tests rather than in a
topically-named split file.

Confidence: high
Scope-risk: narrow
Reversibility: clean
